requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]